proyectodb/requirements.txt
Django==5.1.5
djangorestframework==3.14.0
python-dotenv==1.0.0
sib-api-v3-sdk==7.5.0
mysqlclient==2.2.0
PyMySQL==1.1.0
